extracting the logic from unpublished articles and create artcile form to helpers with custom mades hooks
